refactor: move Toolbar into folder
